+++ src/types/network.rs
@@ -41,3 +41,13 @@ impl std::fmt::Display for RemotePeerId {
 pub struct NetworkId {
     pub(crate) network_id: u16,
 }
+
+// A banned peer identified by its IP address.
+#[derive(Debug)]
+pub struct BannedPeer(pub std::net::IpAddr);
+
+/// A peer to ban identified by its IP address.
+#[derive(Debug)]
+pub enum PeerToBan {
+    IpAddr(IpAddr),
+}
